server/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.1
django-debug-toolbar==2.1
django-extensions==2.2.5
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.8
Pillow==6.2.1
psycopg2==2.8.4
pydot==1.4.1
pyparsing==2.4.5
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.8
